# Optimize `PrevEventIDs` when getting thousands of backwards extremeties