@S7evinK 2024-01-19
Running the benchmark again on a better CPU (allocs etc. stay the same):
```
goos: linux
goarch: amd64
pkg: github.com/matrix-org/dendrite/roomserver/api
cpu: 12th Gen Intel(R) Core(TM) i5-12500H
                               │   old.txt    │               new.txt               │
                               │    sec/op    │   sec/op     vs base                │
PrevEventIDs/Benchmark1-16        127.7n ± 4%   124.6n ± 2%   -2.39% (p=0.011 n=10)
PrevEventIDs/Benchmark2-16        242.0n ± 2%   221.8n ± 5%   -8.35% (p=0.000 n=10)
PrevEventIDs/Benchmark4-16        487.5n ± 4%   432.3n ± 5%  -11.31% (p=0.000 n=10)
PrevEventIDs/Benchmark10-16       1.456µ ± 2%   1.219µ ± 3%  -16.31% (p=0.000 n=10)
PrevEventIDs/Benchmark50-16       12.64µ ± 4%   11.83µ ± 2%   -6.45% (p=0.000 n=10)
PrevEventIDs/Benchmark100-16      28.37µ ± 6%   26.46µ ± 1%   -6.73% (p=0.000 n=10)
PrevEventIDs/Benchmark300-16      96.52µ ± 1%   93.61µ ± 2%   -3.01% (p=0.000 n=10)
PrevEventIDs/Benchmark500-16      173.3µ ± 3%   116.4µ ± 3%  -32.79% (p=0.000 n=10)
PrevEventIDs/Benchmark800-16      291.7µ ± 8%   119.7µ ± 3%  -58.96% (p=0.000 n=10)
PrevEventIDs/Benchmark1000-16     392.9µ ± 5%   118.6µ ± 2%  -69.81% (p=0.000 n=10)
PrevEventIDs/Benchmark2000-16     835.6µ ± 5%   125.8µ ± 4%  -84.95% (p=0.000 n=10)
PrevEventIDs/Benchmark3000-16    1398.8µ ± 1%   141.6µ ± 3%  -89.88% (p=0.000 n=10)
PrevEventIDs/Benchmark5000-16    2632.2µ ± 6%   201.3µ ± 5%  -92.35% (p=0.000 n=10)
PrevEventIDs/Benchmark10000-16   6710.5µ ± 7%   308.9µ ± 4%  -95.40% (p=0.000 n=10)
geomean                           47.65µ        18.97µ       -60.19%
```